it works
like the previous iteration of the generator, when testing on fastapi the req body should be like this

{
  "id": 0,
  "posts": [],
  "noun": "",
  "verb": "",
  "adj": "",
  "emotion": "",
  "style": "",
  "colors": []
}

omitting all nullable params
