+++ backend/services/generator.py
@@ -2,39 +2,52 @@ from ..models import Challenge
 from urllib.request import urlopen
 import random
 import nltk
-# nltk.download('popular')
+import requests
 
-word_site = "https://www.mit.edu/~ecprice/wordlist.10000"
+API_URL = "https://grizzly-fifth-citrine.glitch.me/api"
 
-response = urlopen(word_site)
-txt = response.read().decode()
-WORDS = txt.splitlines()
+def noun() -> str:
+    params = {"generator": "common-noun", "list": "noun"}
+    return requests.get(API_URL, params=params).text
 
-tagged = nltk.pos_tag(WORDS)
+def verb() -> str:
+    params = {"generator": "verb", "list": "verb"}
+    return requests.get(API_URL, params=params).text
 
-nouns = [item[0] for item in tagged if item[1][0] == 'N']
-verbs = [item[0] for item in tagged if item[1][0] == 'V']
-adjs = [item[0] for item in tagged if item[1][0] == 'J']
-styles = ['Impressionism', 'Avant-garde', 'Baroque', 'Cubism', 'Pixel Art', 'Expressionism', 'Futurism','Minimalism','Pop Art','Rococo','Surrealism']
-emotions = ['Sad','Happy','Terror','Hunger']
+def adj() -> str:
+    params = {"generator": "adjective", "list": "adjective"}
+    return requests.get(API_URL, params=params).text
 
-def generate_color(x):
-    colors = []
-    for i in range(0,x):
-        color = random.randrange(0, 2**24)
-        hex_color = hex(color)
-        hexcode = "#" + hex_color[2:]
-        colors.append(hexcode)
-    return colors
+def style() -> str:
+    params = {"generator": "artism", "list": "output"}
+    return requests.get(API_URL, params=params).text
+
+def emotion() -> str:
+    params = {"generator": "emotion", "list": "emotion"}
+    return requests.get(API_URL, params=params).text
+
+def emotion() -> str:
+    params = {"generator": "emotion", "list": "emotion"}
+    return requests.get(API_URL, params=params).text
+
+def colors(number: int | None) -> list:
+    if number == None:
+        return []
+    else:
+        params = {"generator": "hex-code-gen", "list": "output"}
+        result = []
+        for i in range(number):
+            result.append(requests.get(API_URL, params=params).text)
+        return result
 
 def generate(x) -> Challenge:
     result = Challenge(
         id=None,
         posts=[],
-        noun=random.choice(nouns), 
-        verb=random.choice(verbs), 
-        adj=random.choice(adjs), 
-        style=random.choice(styles), 
-        emotion=random.choice(emotions),
-        colors=generate_color(x))
+        noun=noun(), 
+        verb=verb(), 
+        adj=adj(), 
+        style=style(), 
+        emotion=emotion(),
+        colors=colors(x))
     return result
